# src/doc/swagger_v1.yaml
openapi: 3.0.0
info:
  title: Events API
  description: API for managing events
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /events:
    get:
      summary: Get a list of events
      parameters:
        - name: limit
          in: query
          description: Number of events to return
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          description: Number of events to skip
          required: false
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          description: Search term for events
          required: false
          schema:
            type: string
        - name: sort
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [title, description, location, date, price, created_at, updated_at]
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: A list of events
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  meta:
                    type: object
                    properties:
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: integer
                          offset:
                            type: integer
                      order:
                        type: object
                        properties:
                          sort:
                            type: string
                          order:
                            type: string
                      filters:
                        type: object
                        properties:
                          search:
                            type: string
        '404':
          description: No events found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '201':
          description: Event created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
  /events/{id}:
    get:
      summary: Get an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    put:
      summary: Update an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Event'
      responses:
        '200':
          description: Event updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
    delete:
      summary: Delete an event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
components:
  schemas:
    Event:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        date:
          type: string
        price:
          type: number
        image:
          type: string
        userId:
          type: string
        created_at:
          type: string
        updated_at:
          type: string